fix: KPI-161 solve changes in delete scenario service
