+++ manifest.yml
@@ -368,12 +368,13 @@ functions:
     requirements: ./src/handlers/scenarios/requirements.txt
     include:
     - ./src/handlers/scenarios/delete_scenarios_handler.py
-    - ./src/service/scenarios/delete_scenarios_service.py
+    - ./src/service/scenario/delete_scenarios_service.py
     - ./src/utils/query_builder.py
     - ./src/utils/response_sql.py
     - ./db/utils/connection.py
     - ./src/exceptions/base_exception.py
     - ./src/utils/verify_user_permissions.py
+    - ./src/utils/app_names.py    
 
   add_scenario_handler:
     requirements: ./src/handlers/scenario/requirements.txt
