+++ src/tests/config_imports.py
@@ -25,5 +25,4 @@ sys.path.append(f"{root_path}/src/handlers/users")
 sys.path.append(f"{root_path}/src/handlers/comparison_vs_peers")
 sys.path.append(f"{root_path}/src/service/investment_peers_report")
 sys.path.append(f"{root_path}/src/service/by_metric_report")
-sys.path.append(f"{root_path}/src/service/scenarios")
 sys.path.append(f"{root_path}/src/service/edit_modify")
